requirements.txt
anthropic
streamlit
python-dotenv
chromadb
datasets
pandas